+++ src/color.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "stdafx.h"
 
-sf::Color interpolateColor(sf::Color color1, sf::Color color2, double factor);
+sf::Color interpolateColor(sf::Color color1, sf::Color color2, float factor);
