+++ src/stdafx.h
@@ -25,7 +25,7 @@ using std::shared_ptr;
 using std::string;
 
 #define PI 3.1415926535897932384626433832795028841971693993751058209749445923078164062862089986280348253421170679
-#define PIf (double)PI
+#define PIf (float)PI
 
 #define _D(x) std::cout << x << std::endl;
 #define _DV(vec) _D(vec.x << " " << vec.y);
